requirements.txt
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0
pandas>=2.2.0
folium>=0.14.0
streamlit-folium>=0.13.0
python-dotenv>=1.0.0
